--- src/main/java/org/kerminator/hello/HelloController.java
@@ -1,24 +1,0 @@
-package org.kerminator.hello;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class HelloController {
-
-    @RequestMapping("/")
-    public String hello() {
-        return "Hello World";
-    }
-    
-    @RequestMapping("/test")
-    public String test() {
-        return "Test Test";
-    }
-}
