+++ src/main/resources/application.yaml
@@ -1,3 +1,9 @@
+spring:
+  cloud:
+    kubernetes:
+      reload:
+        enabled: true
+
 management:
   endpoints:
     web:
